+++ src/DependencyInjection/HeimrichHannotBackendLostPasswordExtension.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\BackendLostPasswordBundle\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
+
+class HeimrichHannotBackendLostPasswordExtension extends Extension
+{
+    public function load(array $configs, ContainerBuilder $container)
+    {
+        $configuration = new Configuration();
+
+        $config = $this->processConfiguration($configuration, $configs);
+
+        $container->setParameter('huh_backend_lost_password', $config);
+    }
+
+    public function getAlias()
+    {
+        return 'huh_backend_lost_password';
+    }
+}